Add player velocity protocol buffer handler

+++ src/main/scala/server/domain/entities/Player.scala
@@ -29,8 +29,8 @@ object Player {
       initialState: PlayerState
   ) extends Command
   final case class Move(
-      velX: Int,
-      velY: Int,
+      velX: Float,
+      velY: Float,
       replyTo: ActorRef[PlayerHandler.MoveReply]
   ) extends Command
   final case class PrintPosition() extends Command
