+++ src/main/scala/server/domain/structs/PlayerState.scala
@@ -2,6 +2,6 @@ package server.domain.structs
 
 import akka.serialization.jackson.CborSerializable
 
-final case class PlayerPosition(x: Int, y: Int)
+final case class PlayerPosition(x: Float, y: Float)
 
 final case class PlayerState(position: PlayerPosition) extends CborSerializable
